pyproject.toml
[tool.poetry]
name = "custom-openai-api-endpoint"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.7"
langchain-community = "^0.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
